holy flurkin' snit.
Green smoke test.  Did 10K within timeouts.  Lol.

Tighted up state rules and reimplemented Timers - use Duration, etc,
rather than round ticks - lol.

+++ consortium/src/main/java/com/salesforce/apollo/consortium/EnqueuedTransaction.java
@@ -41,7 +41,7 @@ public class EnqueuedTransaction {
         return hash.equals(other.hash);
     }
 
-    public int getDelay() {
+    public long getDelay() {
         Timer current = timer;
         if (current == null) {
             return Integer.MAX_VALUE;
